+++ src/useTable.ts
@@ -347,7 +347,7 @@ export const useTable = <T = any>(rootLocator: Locator, configOptions: TableConf
         
         const columnOverride = config.columnOverrides?.[columnName as keyof T];
         if (columnOverride?.read) {
-            return await columnOverride.read(cell) as R;
+            return await columnOverride.read(cell, { row, columnName, columnIndex: map.get(columnName)! }) as R;
         }
         
         const text = await cell.innerText();